# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "description": "This is the API documentation for your Laravel project.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/www/public",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/exams": {
            "get": {
                "tags": [
                    "Exams"
                ],
                "summary": "List all exams",
                "operationId": "a54e05ebd2b23c9f2329a2a483272bc3",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Exam"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Exams"
                ],
                "summary": "Create an exam",
                "operationId": "fa2de658c3313746441306155597563e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "question_count",
                                    "duration_minutes"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "New Final Exam"
                                    },
                                    "question_count": {
                                        "type": "integer",
                                        "example": 60
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 49.99
                                    },
                                    "link": {
                                        "type": "string",
                                        "example": "https://example.com/exam/page",
                                        "nullable": true
                                    },
                                    "duration_minutes": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "discount_percent": {
                                        "type": "number",
                                        "example": 5,
                                        "nullable": true
                                    },
                                    "quiz_type": {
                                        "type": "string",
                                        "example": "single-choice",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Science summer assessment",
                                        "nullable": true
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "example": "https://example.com/exam.jpg",
                                        "nullable": true
                                    },
                                    "describtion": {
                                        "type": "string",
                                        "example": "For eligible students only",
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Exam created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exam"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/exams/{id}": {
            "get": {
                "tags": [
                    "Exams"
                ],
                "summary": "Get an exam by ID",
                "operationId": "5762ee3221f12d62693826bab5e66e2a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exam"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Exam not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Exams"
                ],
                "summary": "Update an exam",
                "operationId": "d8b5e491c38a96d9bd5f9787cff42ab1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Exam"
                                    },
                                    "question_count": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 70.5
                                    },
                                    "link": {
                                        "type": "string",
                                        "example": "https://example.com/new-link"
                                    },
                                    "duration_minutes": {
                                        "type": "integer",
                                        "example": 80
                                    },
                                    "discount_percent": {
                                        "type": "number",
                                        "example": 0
                                    },
                                    "quiz_type": {
                                        "type": "string",
                                        "example": "essay"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "example": "https://example.com/updated.jpg"
                                    },
                                    "describtion": {
                                        "type": "string",
                                        "example": "Some staff update note"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exam updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exam"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Exam not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Exams"
                ],
                "summary": "Delete an exam",
                "operationId": "61fff9ddd42647f37ba3b48f3c0ac233",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Exam deleted"
                    },
                    "404": {
                        "description": "Exam not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Exam": {
                "required": [
                    "title",
                    "question_count",
                    "duration_minutes"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Mathematics Entrance Test"
                    },
                    "question_count": {
                        "type": "integer",
                        "example": 50
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 99.5,
                        "nullable": true
                    },
                    "link": {
                        "type": "string",
                        "example": "https://example.com/exam/1",
                        "nullable": true
                    },
                    "duration_minutes": {
                        "type": "integer",
                        "example": 120
                    },
                    "discount_percent": {
                        "type": "number",
                        "example": 15,
                        "nullable": true
                    },
                    "quiz_type": {
                        "type": "string",
                        "example": "multiple-choice",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Annual entrance exam",
                        "nullable": true
                    },
                    "image_url": {
                        "type": "string",
                        "example": "https://example.com/q.jpg",
                        "nullable": true
                    },
                    "describtion": {
                        "type": "string",
                        "example": "Extra exam note",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}